feat: add docList composable
